requirements.txt
pandas>=2.2
numpy>=1.26
scikit-learn>=1.4
joblib>=1.4
python-dotenv>=1.0
requests>=2.32
pydantic>=2.8
alpaca-py>=0.21
tenacity>=8.2
loguru>=0.7
tabulate>=0.9
# Institutional-lite add-ons
vectorbt>=0.26
optuna>=3.6
plotly>=5.22
streamlit>=1.35
alpaca-py>=0.20.0
pandas
numpy
scikit-learn
joblib
tenacity
python-dotenv
pydantic
requests
schedule
openpyxl
